chore: move prebuilt configurations to a separate object
Move the prebuilt cache and topic client configurations to their own
Configurations objects, so they don't need to be referenced using
'companion'.

Remove unused example test classes.
