--- src/androidUnitTest/kotlin/software/momento/kotlin/sdk/AndroidExampleTest.kt
@@ -1,3 +1,0 @@
-package software.momento.kotlin.sdk
-
-class AndroidExampleTest
