--- src/commonTest/kotlin/software/momento/kotlin/sdk/CommonExampleTest.kt
@@ -1,2 +1,0 @@
-package software.momento.kotlin.sdk
-class CommonExampleTest
