+++ src/jvmTest/kotlin/software/momento/kotlin/sdk/TopicClientTest.kt
@@ -3,7 +3,7 @@ package software.momento.kotlin.sdk
 import kotlinx.coroutines.test.runTest
 import org.junit.Test
 import software.momento.kotlin.sdk.auth.CredentialProvider
-import software.momento.kotlin.sdk.config.TopicConfiguration
+import software.momento.kotlin.sdk.config.TopicConfigurations
 import software.momento.kotlin.sdk.responses.topic.TopicPublishResponse
 
 
@@ -13,7 +13,7 @@ class TopicClientTest {
     fun testPublish() = runTest {
         val topicClient = TopicClient(
             credentialProvider = CredentialProvider.fromEnvVar("TEST_API_KEY"),
-            configuration = TopicConfiguration.Companion.Laptop.Latest
+            configuration = TopicConfigurations.Laptop.Latest
         )
         val response = topicClient.publish("cache", "topic", "jvm!")
         if (response is TopicPublishResponse.Error) {
